+++ package.json
@@ -15,22 +15,24 @@
     "clean": "gatsby clean"
   },
   "dependencies": {
-    "babel-plugin-styled-components": "^2.0.6",
-    "bootstrap": "^5.2.1",
-    "gatsby": "^4.11.1",
-    "gatsby-plugin-image": "^2.11.1",
-    "gatsby-plugin-manifest": "^4.22.0",
-    "gatsby-plugin-netlify": "^5.0.1",
-    "gatsby-plugin-sharp": "^4.11.1",
-    "gatsby-plugin-styled-components": "^5.11.0",
-    "gatsby-source-filesystem": "^4.11.1",
-    "gatsby-transformer-sharp": "^4.11.0",
-    "react": "^17.0.1",
-    "react-bootstrap": "^2.5.0",
-    "react-dom": "^17.0.1",
+    "babel-plugin-styled-components": "^2.1.4",
+    "gatsby": "^5.11.0",
+    "gatsby-plugin-ffmpeg": "^0.6.4",
+    "gatsby-plugin-image": "^3.11.0",
+    "gatsby-plugin-manifest": "^5.11.0",
+    "gatsby-plugin-netlify": "^5.1.0",
+    "gatsby-plugin-sharp": "^5.11.0",
+    "gatsby-plugin-styled-components": "^6.11.0",
+    "gatsby-source-filesystem": "^5.11.0",
+    "gatsby-transformer-remark": "^6.11.0",
+    "gatsby-transformer-sharp": "^5.11.0",
+    "install": "^0.13.0",
+    "npm": "^9.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-number-format": "^5.0.0-beta.4",
     "react-player": "^2.10.0",
     "react-social-icons": "^5.13.0",
-    "styled-components": "^5.3.5"
+    "styled-components": "^6.0.6"
   }
 }
