+++ src/components/Header.js
@@ -1,39 +1,37 @@
 import React from "react";
 import { Link } from "gatsby";
-import { Navbar, Nav, Container } from "react-bootstrap";
 import { StaticImage } from "gatsby-plugin-image";
 
 const Header = () => {
   return (
     <>
-      <Navbar bg="light" expand="md">
-        <Container>
-          <Navbar.Brand as={Link} to="/" className="d-flex">
-            <StaticImage src="../images/etrLogo.png" alt="etr-logo" width={60} />
+      <div>
+        <div>
+          <Link to="/" className="d-flex">
+            <StaticImage src="..content/images/etrLogo.png" alt="etr-logo" width={60} />
             <span className="p-2">MEDIA PRODUCTIONS</span>
-          </Navbar.Brand>
-          <Navbar.Toggle aria-controls="headerNav" />
-          <Navbar.Collapse id="headerNav">
-            <Nav className="ms-auto">
-              <Nav.Link as={Link} to="/">
+          </Link>
+          <div>
+            <div className="ms-auto">
+              <Link to="/">
                 Home
-              </Nav.Link>
-              <Nav.Link as={Link} to="/services">
+              </Link>
+              <Link to="/services">
                 Services
-              </Nav.Link>
-              <Nav.Link as={Link} to="/portfolio">
+              </Link>
+              <Link to="/portfolio">
                 Our Work
-              </Nav.Link>
-              <Nav.Link as={Link} to="/about">
+              </Link>
+              <Link to="/about">
                 About ETR
-              </Nav.Link>
-              <Nav.Link as={Link} to="/contact">
+              </Link>
+              <Link to="/contact">
                 Contact Us
-              </Nav.Link>
-            </Nav>
-          </Navbar.Collapse>
-        </Container>
-      </Navbar>
+              </Link>
+            </div>
+          </div>
+        </div>
+      </div>
     </>
   );
 };
