+++ src/components/YoutubeVideo.jsx
@@ -1,10 +1,9 @@
 import React from "react";
-import { Ratio } from "react-bootstrap";
 
 const YoutubeVideo = ({ id, height=315, width=560, title }) => {
   const src = 'https://www.youtube.com/embed/' + id + '?modestbranding=1&rel=0'
   return (
-    <Ratio aspectRatio='16x9'>
+    <div>
       <iframe
         width={width}
         height={height}
@@ -14,7 +13,7 @@ const YoutubeVideo = ({ id, height=315, width=560, title }) => {
         allow="accelerometer; autoplay; clipboard-write; encrypted-media;"
         allowfullscreen
       ></iframe>
-    </Ratio>
+    </div>
   );
 };
