+++ src/content/pageData/index.md
@@ -1,0 +1,17 @@
+---
+path: /index
+heading: Show the world your story
+backgroundVideo: ../video/etr-hero-vid.mp4
+tiles:
+  - title: About ETR
+    path: /About
+    image: ../images/filming.jpg
+  - title: Some of Our Work
+    path: /OurWork
+    image: ../images/lense.jpg
+  - title: Get in Touch
+    path: /Contact
+    image: ../images/contact.jpg
+mainpitch: At Empower the Rebel Media Productions, we help individuals and
+  companies showcase their work and products.
+---
